Revert "fixup! shell: xtimerfy sc_icmpv6"
This reverts commit 9f70d5c12f6d52d76adadb5970b1eff2bcc0dc8b.

Conflicts:
	sys/shell/commands/sc_icmpv6_echo.c

+++ sys/shell/commands/sc_icmpv6_echo.c
@@ -152,7 +152,7 @@ int _icmpv6_ping(int argc, char **argv)
     gnrc_netreg_entry_t *ipv6_entry, my_entry = { NULL, ICMPV6_ECHO_REP,
                                                   thread_getpid()
                                                 };
-    uint32_t min_rtt = UINT64_MAX, max_rtt = 0;
+    uint32_t min_rtt = UINT32_MAX, max_rtt = 0;
     uint64_t sum_rtt = 0;
     uint64_t total_time;
     int param_offset = 0;
